reformat

+++ src/main/java/com/jaquadro/minecraft/storagedrawers/item/ItemKey.java
@@ -26,36 +26,36 @@ import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import java.util.List;
 
-public class ItemKey extends Item
-{
+public class ItemKey extends Item {
+
     @CapabilityInject(IDrawerAttributes.class)
     public static Capability<IDrawerAttributes> DRAWER_ATTRIBUTES_CAPABILITY = null;
 
     @Override
     @SideOnly(Side.CLIENT)
-    public void addInformation (@Nonnull ItemStack itemStack, @Nullable World world, List<String> list, ITooltipFlag advanced) {
+    public void addInformation(@Nonnull ItemStack itemStack, @Nullable World world, List<String> list, ITooltipFlag advanced) {
         String name = getTranslationKey(itemStack);
         list.add(I18n.format(name + ".description"));
     }
 
     @SideOnly(Side.CLIENT)
-    public boolean isFull3D () {
+    public boolean isFull3D() {
         return true;
     }
 
     @Override
     @Nonnull
-    public Multimap<String, AttributeModifier> getAttributeModifiers (EntityEquipmentSlot slot, @Nonnull ItemStack stack) {
+    public Multimap<String, AttributeModifier> getAttributeModifiers(EntityEquipmentSlot slot, @Nonnull ItemStack stack) {
         Multimap<String, AttributeModifier> multimap = super.getAttributeModifiers(slot, stack);
 
         if (slot == EntityEquipmentSlot.MAINHAND)
-            multimap.put(SharedMonsterAttributes.ATTACK_DAMAGE.getName(), new AttributeModifier(ATTACK_DAMAGE_MODIFIER, "Weapon modifier", (double)2, 0));
+            multimap.put(SharedMonsterAttributes.ATTACK_DAMAGE.getName(), new AttributeModifier(ATTACK_DAMAGE_MODIFIER, "Weapon modifier", 2, 0));
 
         return multimap;
     }
 
     @Override
-    public EnumActionResult onItemUse (EntityPlayer player, World world, BlockPos pos, EnumHand hand, EnumFacing facing, float hitX, float hitY, float hitZ) {
+    public EnumActionResult onItemUse(EntityPlayer player, World world, BlockPos pos, EnumHand hand, EnumFacing facing, float hitX, float hitY, float hitZ) {
         TileEntity tile = world.getTileEntity(pos);
         if (tile == null)
             return EnumActionResult.PASS;
@@ -69,5 +69,6 @@ public class ItemKey extends Item
         return EnumActionResult.SUCCESS;
     }
 
-    protected void handleDrawerAttributes (IDrawerAttributesModifiable attrs) { }
+    protected void handleDrawerAttributes(IDrawerAttributesModifiable attrs) {
+    }
 }
