+++ src/main/java/com/jaquadro/minecraft/storagedrawers/item/ItemPersonalKey.java
@@ -17,9 +17,9 @@ import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import java.util.List;
 
-public class ItemPersonalKey extends Item
-{
-    public ItemPersonalKey (String registryName, String unlocalizedName) {
+public class ItemPersonalKey extends Item {
+
+    public ItemPersonalKey(String registryName, String unlocalizedName) {
         setRegistryName(registryName);
         setTranslationKey(unlocalizedName);
         setCreativeTab(ModCreativeTabs.tabStorageDrawers);
@@ -28,28 +28,27 @@ public class ItemPersonalKey extends Item
 
     @Override
     @SideOnly(Side.CLIENT)
-    public void addInformation (@Nonnull ItemStack itemStack, @Nullable World world, List<String> list, ITooltipFlag advanced) {
+    public void addInformation(@Nonnull ItemStack itemStack, @Nullable World world, List<String> list, ITooltipFlag advanced) {
         String name = getTranslationKey(itemStack);
         list.add(I18n.format(name + ".description"));
     }
 
     @SideOnly(Side.CLIENT)
-    public boolean isFull3D()
-    {
+    public boolean isFull3D() {
         return true;
     }
 
     @Override
-    public Multimap<String, AttributeModifier> getAttributeModifiers (EntityEquipmentSlot slot, @Nonnull ItemStack stack) {
+    public Multimap<String, AttributeModifier> getAttributeModifiers(EntityEquipmentSlot slot, @Nonnull ItemStack stack) {
         Multimap<String, AttributeModifier> multimap = super.getAttributeModifiers(slot, stack);
 
         if (slot == EntityEquipmentSlot.MAINHAND)
-            multimap.put(SharedMonsterAttributes.ATTACK_DAMAGE.getName(), new AttributeModifier(ATTACK_DAMAGE_MODIFIER, "Weapon modifier", (double)2, 0));
+            multimap.put(SharedMonsterAttributes.ATTACK_DAMAGE.getName(), new AttributeModifier(ATTACK_DAMAGE_MODIFIER, "Weapon modifier", 2, 0));
 
         return multimap;
     }
 
-    public String getSecurityProviderKey (int meta) {
+    public String getSecurityProviderKey(int meta) {
         return null;
     }
 }
